# 11-project-ai-prompt-management/docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.AnalyticsLog:
    properties:
      api_key_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      latency_ms:
        type: integer
      prompt_id:
        type: integer
      response_code:
        type: integer
      token_estimate:
        type: integer
    type: object
  entity.ApiKey:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      masked_key:
        description: prompt_live_xxxx...1234
        type: string
      name:
        type: string
      workspace_id:
        type: integer
    type: object
...
  entity.Workspace:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.CreateApiKeyReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  handler.CreateWorkspaceReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a production-grade multi-tenant AI Prompt Management API with
    Redis caching and offline JWT verification.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AI Prompt Management API
  version: "1.0"
paths:
...
  /workspaces:
    post:
      consumes:
      - application/json
      description: Creates a new workspace. The creator automatically becomes the
        workspace Admin. Requires JWT authorization.
      parameters:
      - description: Workspace name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateWorkspaceReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Workspace'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new workspace
      tags:
      - Workspaces
  /workspaces/{id}/analytics:
    get:
      consumes:
      - application/json
      description: Retrieves logs of prompt compiler executions (latency, token estimates).
        Requires workspace Admin/Member JWT authorization.
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.AnalyticsLog'
            type: array
        "400":
          description: Invalid workspace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get workspace execution analytics logs
      tags:
      - Analytics
  /workspaces/{id}/api-keys:
    get:
      consumes:
      - application/json
      description: Lists all API Keys generated for a workspace (masked for security).
        Requires workspace Admin JWT authorization.
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ApiKey'
            type: array
        "400":
          description: Invalid workspace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get workspace API Keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Generates a new API Key for client integration. The key string
        is only shown once in the response. Requires workspace Admin JWT authorization.
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: integer
      - description: Key descriptor name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateApiKeyReq'
      produces:
      - application/json
      responses:
        "201":
          description: Returns raw api_key, masked_key, id, and expires_at
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden access to workspace
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create API Key for workspace
      tags:
      - API Keys
  /workspaces/{id}/api-keys/{key_id}:
    delete:
      consumes:
      - application/json
      description: Deletes and revokes an API Key, invalidating it from cache instantly.
        Requires workspace Admin JWT authorization.
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: integer
      - description: API Key ID
        in: path
        name: key_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API Key revoked successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input IDs
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Revoke an API Key
      tags:
      - API Keys
...
securityDefinitions:
  ApiKeyAuth:
    description: Paste your JWT token in the format "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
  ClientApiKeyAuth:
    description: Paste your API Key in the format "Bearer prompt_live_<key>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"